Move solana test client to solana workspace
both packet verifiers share a test impl
